+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "start": "ts-node ./index.ts",
+    "start": "rm -rf ./dist && tsc && node ./dist/index.js",
+    "start:debug": "tsc &&  node --trace-warnings ./dist/index.js",
     "start:dev": "nodemon ./index.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
