+++ src/positioning.ts
@@ -1,7 +1,7 @@
 import { Duplex } from 'stream';
 import robot from 'robotjs';
-import Point from './point.js';
-import { DIRECTION, COMMAND } from './constants.js';
+import Point from './drawer/point.js';
+import { DIRECTION, COMMAND } from './common/constants.js';
 import { sendCommandWithLog } from './command.js';
 
 export async function sendPosition(writeStream: Duplex): Promise<void> {
